+++ engines/colony/renderer.h
@@ -59,7 +59,7 @@ public:
 
 	// Buffer management
 	virtual void copyToScreen() = 0;
-	virtual void setWireframe(bool enable, int fillColor = -1) = 0;
+	virtual void setWireframe(bool enable, int64_t fillColor = -1) = 0;
 	virtual void setStippleData(const byte *data) {}
 	virtual void setMacColors(uint32 fg, uint32 bg) {}
 	virtual void computeScreenViewport() = 0;
